test(integration): move to the test directory
