+++ test/integration/data/registry/auth/htpasswd
