+++ test/integration/data/registry/certs/cert.pem
