+++ test/integration/data/registry/certs/key.pem
