+++ test/integration/testdata/goldens/packages/alpine-3.10.json.golden
