+++ test/integration/testdata/goldens/packages/amazonlinux-2.json.golden
