+++ test/integration/testdata/goldens/packages/centos-6.json.golden
