+++ test/integration/testdata/goldens/packages/gcr.io-distroless-base-latest.json.golden
