+++ test/integration/testdata/goldens/packages/knqyf263-vuln-image-1.2.3.json.golden
