+++ test/integration/testdata/goldens/packages/opensuse-leap-latest.json.golden
