+++ test/integration/testdata/goldens/packages/oraclelinux-8-slim.json.golden
