+++ test/integration/testdata/goldens/packages/photon-1.0-20190823.json.golden
