+++ test/integration/testdata/goldens/packages/registry.redhat.io-ubi7.json.golden
