+++ test/integration/testdata/goldens/packages/ubuntu-18.04.json.golden
