+++ test/integration/testdata/goldens/vuln-image1.2.3.expectedlibs.golden
